Made new changes
I have added a code under the rug and
animated the code under rug. I also scripted the code to move out when E is pressed and camera moves closer to the code. I have added blue code invisible when light is on. I have scripted the blue code to be on and off when light is off and on.