Made new changes
I finished creating Room 2 and decided which puzzles will be in this room. I also added camera zoom to the clock and controls to change the hours needle and also the minutes needle. I added code for if  the correct time is activated the key will appear and we can collect the key.